requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0
exchangelib>=4.9.0
pytz>=2021.3
ollama>=0.2.0
python-dotenv>=0.19
pyinstaller>=5.0
apscheduler==3.10.4
gTTS